+++ packages/core/src/index.ts
@@ -59,6 +59,7 @@ export type {
   MetaSession,
   MetaUser,
   MetaView,
+  MetaOverrides,
 } from './metas';
 
 export {